+++ landing-page/components/sections/landing/ExpansionSection.tsx
