+++ landing-page/components/sections/landing/Partners.tsx
