+++ landing-page/components/sections/landing/Story.tsx
